requirements.txt
langchain
wikipedia
langchain-community
langchain_groq
python-dotenv
pydantic
pyttsx3